No warnings

+++ ProjectSFML/Zengine/Environment/Ground.cpp
@@ -23,9 +23,9 @@ void Ground::SetBoxColliders(string pathToGroundTxt)
 		file >> sizeX;
 		file >> sizeY;
 
-		zenObjects.push_back(new ZenObject(std::stoi(id), "Ground", sf::Vector2f(std::stoi(sizeX), std::stoi(sizeY))));
+		zenObjects.push_back(new ZenObject(std::stoi(id), "Ground", sf::Vector2f((float)(std::stoi(sizeX)), (float)(std::stoi(sizeY)))));
 
-		boxColliders.push_back(new BoxCollider2D(Vector2(std::stoi(positionX), std::stoi(positionY)), Vector2(std::stoi(sizeX), std::stoi(sizeY)), zenObjects[i], Collider::ColliderTags::GROUND));
+		boxColliders.push_back(new BoxCollider2D(Vector2((float)(std::stoi(positionX)), (float)(std::stoi(positionY))), Vector2((float)(std::stoi(sizeX)), (float)(std::stoi(sizeY))), zenObjects[i], Collider::ColliderTags::GROUND));
 	}
 
 	file.close();
